+++ domain/business/src/main/java/com/navektest/business/search/SearchMulti.kt
@@ -25,7 +25,7 @@ data class SearchMulti(
         val release_date: String,
         val title: String,
         val video: Boolean,
-        val vote_average: Int,
+        val vote_average: Double,
         val vote_count: Int
     )
